Improve the logic of user to handle better the dates

+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository, Between } from 'typeorm';
 import * as bcrypt from 'bcrypt';
@@ -9,6 +9,7 @@ import { LoginUserDto } from './dto/login-user.dto';
 import { ChangePasswordDto } from './dto/change-password.dto';
 import { Sale } from '../sales/entities/sale.entity';
 import { AccountsHoldings } from 'src/accountsholdings/entities/accountsholding.entity';
+import * as moment from 'moment-timezone';
 
 @Injectable()
 export class UsersService {
@@ -63,14 +64,25 @@ export class UsersService {
   }
 
   async updateUser(id: number, updateUserDto: UpdateUserDto) {
+
+    const userExists = await this.userRepository.findOne({ where: { id: id } });
+
+    if (!userExists) { throw new HttpException('User not found', HttpStatus.NOT_FOUND); }
+
     const user = await this.getUserById(id);
     Object.assign(user, updateUserDto);
     return this.userRepository.save(user);
   }
 
   async deleteUser(id: number) {
+
+    const userExists = await this.userRepository.findOne({ where: { id: id } });
+
+    if (!userExists) { throw new HttpException('User not found', HttpStatus.NOT_FOUND); }
+
     const user = await this.getUserById(id);
     return this.userRepository.remove(user);
+
   }
 
   async verifyPassword(userId: number, password: string) {
@@ -79,24 +91,55 @@ export class UsersService {
   }
 
   async changePassword(changePasswordDto: ChangePasswordDto) {
+
+    const userExists = await this.userRepository.findOne({ where: { id: changePasswordDto.userId } });
+
+    if (!userExists) { throw new HttpException('User not found', HttpStatus.NOT_FOUND); }
+
     const user = await this.getUserById(changePasswordDto.userId);
     user.password = await bcrypt.hash(changePasswordDto.newPassword, 10);
     return this.userRepository.save(user);
   }
 
-  async getMonthlySalesByUser(userId: number, year: number) {
-    const result = await this.saleRepository
-      .createQueryBuilder('sale')
-      .select('MONTH(sale.date)', 'month')
-      .addSelect('SUM(sale.total)', 'totalSales')
-      .where('YEAR(sale.date) = :year AND sale.userId = :userId', { year, userId })
-      .groupBy('month')
-      .orderBy('month', 'ASC')
-      .getRawMany();
-    return result;
+  async getMonthlySalesByUser(userId: number, year: number): Promise<{ month: number; totalSales: number }[]> {
+
+    const user = await this.userRepository.findOne({where: {id: userId}});
+
+    if (!user) {throw new HttpException('User not found', HttpStatus.NOT_FOUND);}
+
+    const timezone = 'America/El_Salvador'; // Zona horaria
+    const results = [];
+  
+    // Iterar sobre cada mes del año
+    for (let month = 0; month < 12; month++) {
+      // Obtener el inicio y el fin del mes en la zona horaria especificada
+      const monthStart = moment.tz({ year, month, day: 1 }, timezone).startOf('month').toDate();
+      const monthEnd = moment.tz({ year, month, day: 1 }, timezone).endOf('month').toDate();
+  
+      // Consultar las ventas del mes para el usuario específico
+      const result = await this.saleRepository
+        .createQueryBuilder('sale')
+        .select('SUM(sale.total)', 'totalSales')
+        .where('sale.date >= :monthStart AND sale.date <= :monthEnd', { monthStart, monthEnd })
+        .andWhere('sale.userId = :userId', { userId })
+        .getRawOne();
+  
+      // Guardar el resultado
+      results.push({
+        month: month + 1, // Los meses en JavaScript van de 0 a 11, por eso sumamos 1
+        totalSales: parseFloat(result.totalSales) || 0, // Si no hay ventas, devolvemos 0
+      });
+    }
+  
+    return results;
   }
 
   async getTotalIncomeByUser(userId: number) {
+
+    const user = await this.userRepository.findOne({where: {id: userId}});
+
+    if (!user) {throw new HttpException('User not found', HttpStatus.NOT_FOUND);}
+
     const salesIncome = await this.saleRepository
       .createQueryBuilder('sale')
       .select('SUM(sale.total)', 'totalIncome')
@@ -113,53 +156,104 @@ export class UsersService {
     return { totalIncome };
   }
 
-  async getTodayIncomeByUser(userId: number) {
+  async getTodayIncomeByUser(userId: number): Promise<{ totalIncome: number }> {
+    const timezone = 'America/El_Salvador'; // Zona horaria
+  
+    // Obtener el usuario
+    const user = await this.userRepository.findOne({ where: { id: userId } });
+    if (!user) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+  
+    // Obtener el inicio y el fin del día en la zona horaria especificada
+    const todayStart = moment().tz(timezone).startOf('day').toDate(); // Inicio del día (00:00:00)
+    const todayEnd = moment().tz(timezone).endOf('day').toDate(); // Fin del día (23:59:59)
+  
+    // Consultar las ventas del día para el usuario específico
     const salesIncome = await this.saleRepository
       .createQueryBuilder('sale')
       .select('SUM(sale.total)', 'totalIncome')
-      .where('DATE(sale.date) = CURDATE() AND sale.userId = :userId', { userId })
+      .where('sale.date >= :todayStart AND sale.date <= :todayEnd', { todayStart, todayEnd })
+      .andWhere('sale.userId = :userId', { userId })
       .getRawOne();
-
+  
+    // Consultar los pagos de cuentas del día para el usuario específico
     const accountsIncome = await this.accountsHoldingsRepository
       .createQueryBuilder('account')
       .select('SUM(account.paid)', 'totalIncome')
-      .where('DATE(account.date) = CURDATE() AND account.userId = :userId', { userId })
+      .where('account.date >= :todayStart AND account.date <= :todayEnd', { todayStart, todayEnd })
+      .andWhere('account.userId = :userId', { userId })
       .getRawOne();
-
+  
+    // Calcular el ingreso total
     const totalIncome = (parseFloat(salesIncome.totalIncome) || 0) + (parseFloat(accountsIncome.totalIncome) || 0);
     return { totalIncome };
   }
 
-  async getWeeklyIncomeByUser(userId: number) {
+  async getWeeklyIncomeByUser(userId: number): Promise<{ totalIncome: number }> {
+
+        // Obtener el usuario
+    const user = await this.userRepository.findOne({ where: { id: userId } });
+
+    if (!user) { throw new HttpException('User not found', HttpStatus.NOT_FOUND); }
+
+    const timezone = 'America/El_Salvador'; // Zona horaria
+  
+    // Obtener el inicio y el fin de la semana actual en la zona horaria especificada
+    const weekStart = moment().tz(timezone).startOf('week').toDate(); // Inicio de la semana (domingo a las 00:00:00)
+    const weekEnd = moment().tz(timezone).endOf('week').toDate(); // Fin de la semana (sábado a las 23:59:59)
+  
+    // Consultar las ventas de la semana para el usuario específico
     const salesIncome = await this.saleRepository
       .createQueryBuilder('sale')
       .select('SUM(sale.total)', 'totalIncome')
-      .where('WEEK(sale.date) = WEEK(NOW()) AND sale.userId = :userId', { userId })
+      .where('sale.date >= :weekStart AND sale.date <= :weekEnd', { weekStart, weekEnd })
+      .andWhere('sale.userId = :userId', { userId })
       .getRawOne();
-
+  
+    // Consultar los pagos de cuentas de la semana para el usuario específico
     const accountsIncome = await this.accountsHoldingsRepository
       .createQueryBuilder('account')
       .select('SUM(account.paid)', 'totalIncome')
-      .where('WEEK(account.date) = WEEK(NOW()) AND account.userId = :userId', { userId })
+      .where('account.date >= :weekStart AND account.date <= :weekEnd', { weekStart, weekEnd })
+      .andWhere('account.userId = :userId', { userId })
       .getRawOne();
-
+  
+    // Calcular el ingreso total
     const totalIncome = (parseFloat(salesIncome.totalIncome) || 0) + (parseFloat(accountsIncome.totalIncome) || 0);
     return { totalIncome };
   }
 
-  async getMonthlyIncomeByUser(userId: number) {
+  async getMonthlyIncomeByUser(userId: number): Promise<{ totalIncome: number }> {
+
+    // Obtener el usuario
+    const user = await this.userRepository.findOne({ where: { id: userId } });
+
+    if (!user) { throw new HttpException('User not found', HttpStatus.NOT_FOUND); }
+
+    const timezone = 'America/El_Salvador'; // Zona horaria
+  
+    // Obtener el inicio y el fin del mes actual en la zona horaria especificada
+    const monthStart = moment().tz(timezone).startOf('month').toDate(); // Inicio del mes (día 1 a las 00:00:00)
+    const monthEnd = moment().tz(timezone).endOf('month').toDate(); // Fin del mes (último día a las 23:59:59)
+  
+    // Consultar las ventas del mes para el usuario específico
     const salesIncome = await this.saleRepository
       .createQueryBuilder('sale')
       .select('SUM(sale.total)', 'totalIncome')
-      .where('MONTH(sale.date) = MONTH(CURRENT_DATE()) AND YEAR(sale.date) = YEAR(CURRENT_DATE()) AND sale.userId = :userId', { userId })
+      .where('sale.date >= :monthStart AND sale.date <= :monthEnd', { monthStart, monthEnd })
+      .andWhere('sale.userId = :userId', { userId })
       .getRawOne();
-
+  
+    // Consultar los pagos de cuentas del mes para el usuario específico
     const accountsIncome = await this.accountsHoldingsRepository
       .createQueryBuilder('account')
       .select('SUM(account.paid)', 'totalIncome')
-      .where('MONTH(account.date) = MONTH(CURRENT_DATE()) AND YEAR(account.date) = YEAR(CURRENT_DATE()) AND account.userId = :userId', { userId })
+      .where('account.date >= :monthStart AND account.date <= :monthEnd', { monthStart, monthEnd })
+      .andWhere('account.userId = :userId', { userId })
       .getRawOne();
-
+  
+    // Calcular el ingreso total
     const totalIncome = (parseFloat(salesIncome.totalIncome) || 0) + (parseFloat(accountsIncome.totalIncome) || 0);
     return { totalIncome };
   }
